feat: Customer 내부 변수 원시값 포장 및 패키지 구조 변경
