+++ src/main/java/woowacourse/shoppingcart/domain/customer/Email.java
@@ -1,0 +1,26 @@
+package woowacourse.shoppingcart.domain.customer;
+
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
+public class Email {
+
+    private static final Pattern EMAIL_PATTERN = Pattern.compile("^[a-z0-9._-]+@[a-z]+[.]+[a-z]{2,3}$");
+
+    private final String value;
+
+    public Email(String value) {
+        this.value = value;
+    }
+
+    private void validateEmail(String email) {
+        Matcher matcher = EMAIL_PATTERN.matcher(email);
+        if (!matcher.matches()) {
+            throw new IllegalArgumentException("email 형식이 올바르지 않습니다. (형식: example@email.com");
+        }
+    }
+
+    public String getValue() {
+        return value;
+    }
+}
