+++ src/test/java/woowacourse/shoppingcart/domain/customer/CustomerTest.java
@@ -1,10 +1,4 @@
-package woowacourse.shoppingcart.domain;
-
-import static org.assertj.core.api.Assertions.*;
-import static org.junit.jupiter.api.Assertions.*;
-import static woowacourse.shoppingcart.CustomerFixtures.*;
-
-import java.util.stream.Stream;
+package woowacourse.shoppingcart.domain.customer;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -12,6 +6,14 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.junit.jupiter.params.provider.ValueSource;
+
+import java.util.stream.Stream;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.junit.jupiter.api.Assertions.assertAll;
+import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
+import static woowacourse.shoppingcart.CustomerFixtures.*;
 
 class CustomerTest {
